Created a Prototype for the Login page
Created a Login page's Prototype using the 'Balsamiq' software tool for the wholesale system by using different features of the software, the login page consists of the login button, links for the Forgot password and the registration. And blocks for the Username and the Password. And title of the web application quoting "Wholesale System".