Updated the existing Login page (xml file)
I have added the login title, and the logo depicting the Wholesale System. I have changed the Email to the username and added symbols for the username and password. At the end of the page added a simple note to create a new registration if the user does not have an existing one.